+++ examples/local_evaluation/huggingface_model_evaluation.py
@@ -1,6 +1,7 @@
+from functools import partial
 from pathlib import Path
 
-from eval_framework.llm.huggingface_llm import HFLLM
+from eval_framework.llm.huggingface import HFLLM
 from eval_framework.main import main
 from eval_framework.tasks.eval_config import EvalConfig
 from template_formatting.formatter import HFFormatter
@@ -9,7 +10,7 @@ from template_formatting.formatter import HFFormatter
 # Define your model
 class MyHuggingFaceModel(HFLLM):
     LLM_NAME = "microsoft/DialoGPT-medium"
-    DEFAULT_FORMATTER = HFFormatter("microsoft/DialoGPT-medium")
+    DEFAULT_FORMATTER = partial(HFFormatter, "microsoft/DialoGPT-medium")
 
 
 if __name__ == "__main__":